refactor: folder structure, create pip package
