+++ latent_flow/datasets/Falor3D.py
@@ -1,26 +1,23 @@
 import os
 import numpy as np
-import torch
 from torch.utils.data import Dataset
 import glob
 import PIL.Image
 
-factor_sizes = [3, 8, 5, 4, 4, 4, 6, 4, 4]
+factor_sizes = [5, 6, 6, 6, 6, 6, 6]
 factor_names = [
-    "object_shape",
-    "robot_x-move",
-    "robot_y-move",
-    "camera_height",
-    "object_scale",
     "lighting_intensity",
+    "lighting_x-dir",
     "lighting_y-dir",
-    "object_color",
-    "wall_color",
+    "lighting_z-dir",
+    "camera_x-pos",
+    "camera_y-pos",
+    "camera_z-pos",
 ]
 
 
-class Isaac3D(Dataset):
-    """Isaac 3D dataset."""
+class Falor3D(Dataset):
+    """Falor3D dataset."""
 
     def __init__(self, root, train=True, transform=None):
         self.image_files = sorted(glob.glob(os.path.join(root, "*.png")))
@@ -32,8 +29,8 @@ class Isaac3D(Dataset):
         return len(self.image_files)
 
     def __getitem__(self, idx):
-        # index = np.random.randint(8) + 1
-        index = self.index + 1
+        # index = np.random.randint(7)
+        index = self.index
         factor_size = factor_sizes[index]
         index_arr = self.factor_bases[index] * np.arange(factor_size) + (idx % self.factor_bases[0])
         image_names = [self.image_files[t] for t in index_arr]
@@ -45,4 +42,5 @@ class Isaac3D(Dataset):
             if self.transform:
                 img = self.transform(img)
             images.append(img)
-        return images, index - 1
+        # images = np.array(images)
+        return images, index
