+++ latent_flow/datasets/Isaac3D.py
@@ -1,24 +1,25 @@
 import os
 import numpy as np
-import torch
 from torch.utils.data import Dataset
 import glob
 import PIL.Image
 
-factor_sizes = [5, 6, 6, 6, 6, 6, 6]
+factor_sizes = [3, 8, 5, 4, 4, 4, 6, 4, 4]
 factor_names = [
+    "object_shape",
+    "robot_x-move",
+    "robot_y-move",
+    "camera_height",
+    "object_scale",
     "lighting_intensity",
-    "lighting_x-dir",
     "lighting_y-dir",
-    "lighting_z-dir",
-    "camera_x-pos",
-    "camera_y-pos",
-    "camera_z-pos",
+    "object_color",
+    "wall_color",
 ]
 
 
-class Falor3D(Dataset):
-    """Falor3D dataset."""
+class Isaac3D(Dataset):
+    """Isaac 3D dataset."""
 
     def __init__(self, root, train=True, transform=None):
         self.image_files = sorted(glob.glob(os.path.join(root, "*.png")))
@@ -30,8 +31,8 @@ class Falor3D(Dataset):
         return len(self.image_files)
 
     def __getitem__(self, idx):
-        # index = np.random.randint(7)
-        index = self.index
+        # index = np.random.randint(8) + 1
+        index = self.index + 1
         factor_size = factor_sizes[index]
         index_arr = self.factor_bases[index] * np.arange(factor_size) + (idx % self.factor_bases[0])
         image_names = [self.image_files[t] for t in index_arr]
@@ -43,5 +44,4 @@ class Falor3D(Dataset):
             if self.transform:
                 img = self.transform(img)
             images.append(img)
-        # images = np.array(images)
-        return images, index
+        return images, index - 1
