+++ latent_flow/datasets/Shapes3d.py
