+++ latent_flow/trainers/trainer_ot_scratch_weakly_shapes.py
@@ -13,21 +13,21 @@ from torch.utils.data import DataLoader
 from torch.utils.tensorboard import SummaryWriter
 from tensorboard import program
 from .aux import sample_z, TrainingStatTracker, update_progress, update_stdout, sec2dhms
-from transforms import *
+
+# from transforms import *
 from torch.distributions.normal import Normal
 from torch.autograd import grad
-from lib.DiffPDE import DiffPDE
-from lib.HJPDE import HJPDE
-from vae import ConvVAE
-
-torch.autograd.set_detect_anomaly(True)
 
 
-angle_set = [0, 10, 20, 30, 40, 50, 60, 70, 80]
-color_set = [180, 200, 220, 240, 260, 280, 300, 320, 340]
-scale_set = [1.0, 1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7, 1.8]
-mnist_trans = AddRandomTransformationDims(angle_set=angle_set, color_set=color_set, scale_set=scale_set)
-mnist_color = To_Color()
+def idx2onehot(idx, n):
+    assert torch.max(idx).item() < n
+
+    if idx.dim() == 1:
+        idx = idx.unsqueeze(1)
+    onehot = torch.zeros(idx.size(0), n).to(idx.device)
+    onehot.scatter_(1, idx, 1)
+
+    return onehot
 
 
 def log_normal_diag(x, mean, log_var, average=False, dim=None):
@@ -43,8 +43,16 @@ class DataParallelPassthrough(nn.DataParallel):
             return getattr(self.module, name)
 
 
-class TrainerOTScratch(object):
-    def __init__(self, params=None, exp_dir=None, use_cuda=False, multi_gpu=False, data_loader=None):
+class TrainerOTScratchWeaklyShapes(object):
+    def __init__(
+        self,
+        params=None,
+        exp_dir=None,
+        use_cuda=False,
+        multi_gpu=False,
+        data_loader=None,
+        dataset=None,
+    ):
         if params is None:
             raise ValueError("Cannot build a Trainer instance with empty params: params={}".format(params))
         else:
@@ -52,7 +60,7 @@ class TrainerOTScratch(object):
         self.use_cuda = use_cuda
         self.multi_gpu = multi_gpu
         self.data_loader = data_loader
-
+        self.dataset = dataset
         # Use TensorBoard
         self.tensorboard = self.params.tensorboard
 
@@ -88,35 +96,47 @@ class TrainerOTScratch(object):
         self.cross_entropy = nn.CrossEntropyLoss()
         # Define KL Div
         self.kl_div = nn.KLDivLoss(reduction="batchmean", log_target=True)
+        self.kl_index = nn.KLDivLoss(reduction="batchmean", log_target=False)
         # Array of iteration times
         self.iter_times = np.array([])
 
         # Set up training statistics tracker
         self.stat_tracker = TrainingStatTracker()
 
-    def get_starting_iteration(self, support_sets, generator, prior):
+    def get_starting_iteration(self, support_sets, reconstructor, generator, prior):
+        """Check if checkpoint file exists (under `self.models_dir`) and set starting iteration at the checkpoint
+        iteration; also load checkpoint weights to `support_sets` and `reconstructor`. Otherwise, set starting
+        iteration to 1 in order to train from scratch.
+
+        Returns:
+            starting_iter (int): starting iteration
+
+        """
         starting_iter = 1
         if osp.isfile(self.checkpoint):
             checkpoint_dict = torch.load(self.checkpoint)
             starting_iter = checkpoint_dict["iter"]
             support_sets.load_state_dict(checkpoint_dict["support_sets"])
             prior.load_state_dict(checkpoint_dict["prior"])
-            # state_dict_new = {}
-            # for k, v in checkpoint_dict['vae'].items():
-            #    state_dict_new[k[len("module.encoder."):]] = v
+            reconstructor.load_state_dict(checkpoint_dict["reconstructor"])
+
             generator.load_state_dict(checkpoint_dict["vae"])
         return starting_iter
 
     def loss_fn(self, recon_x, x, mean, log_var):
         BCE = torch.nn.functional.binary_cross_entropy(
-            recon_x.view(self.params.batch_size, -1), x.view(self.params.batch_size, -1), reduction="sum"
+            recon_x.view(self.params.batch_size, -1),
+            x.view(self.params.batch_size, -1),
+            reduction="sum",
         )
         KLD = -0.5 * torch.sum(1 + log_var - mean.pow(2) - log_var.exp())
         return (BCE + KLD) / x.size(0)
 
     def bce(self, recon_x, x):
         BCE = torch.nn.functional.binary_cross_entropy(
-            recon_x.view(self.params.batch_size, -1), x.view(self.params.batch_size, -1), reduction="sum"
+            recon_x.view(self.params.batch_size, -1),
+            x.view(self.params.batch_size, -1),
+            reduction="sum",
         )
         return BCE / x.size(0)
 
@@ -167,7 +187,15 @@ class TrainerOTScratch(object):
         print("         ===================================================================")
         update_stdout(10)
 
-    def train(self, generator: ConvVAE, support_sets: HJPDE, prior: DiffPDE):
+    def train(self, generator, support_sets, reconstructor, prior):
+        """Training function.
+
+        Args:
+            generator     :
+            support_sets  :
+            reconstructor :
+
+        """
         # Save initial `support_sets` model as `support_sets_init.pt`
         torch.save(support_sets.state_dict(), osp.join(self.models_dir, "support_sets_init.pt"))
 
@@ -176,27 +204,34 @@ class TrainerOTScratch(object):
         if self.use_cuda:
             generator.cuda().train()
             support_sets.cuda().train()
+            reconstructor.cuda().train()
             prior.cuda().train()
         else:
             generator.train()
             support_sets.train()
+            reconstructor.train()
             prior.train()
 
         # Set support sets optimizer
         support_sets_optim = torch.optim.Adam(
-            [{"params": support_sets.parameters()}, {"params": prior.parameters()}], lr=self.params.support_set_lr
+            [{"params": support_sets.parameters()}, {"params": prior.parameters()}],
+            lr=self.params.support_set_lr,
         )
 
         # Set VAE optimizer
-        vae_optimizer = torch.optim.Adam(generator.parameters(), lr=self.params.generator_lr)
+        vae_optimizer = torch.optim.Adam(generator.parameters(), lr=self.params.reconstructor_lr)
+
+        # Set shift predictor optimizer
+        reconstructor_optim = torch.optim.Adam(reconstructor.parameters(), lr=self.params.reconstructor_lr)
 
         # Get starting iteration
-        starting_iter = self.get_starting_iteration(support_sets, generator, prior)
+        starting_iter = self.get_starting_iteration(support_sets, reconstructor, generator, prior)
 
         # Parallelize `generator` and `reconstructor` into multiple GPUs, if available and `multi_gpu=True`.
         if self.multi_gpu:
             print("#. Parallelize G, R over {} GPUs...".format(torch.cuda.device_count()))
             generator = DataParallelPassthrough(generator)
+            reconstructor = DataParallelPassthrough(reconstructor)
             cudnn.benchmark = True
 
         # Check starting iteration
@@ -204,7 +239,11 @@ class TrainerOTScratch(object):
             print("#. This experiment has already been completed and can be found @ {}".format(self.wip_dir))
             print("#. Copy {} to {}...".format(self.wip_dir, self.complete_dir))
             try:
-                shutil.copytree(src=self.wip_dir, dst=self.complete_dir, ignore=shutil.ignore_patterns("checkpoint.pt"))
+                shutil.copytree(
+                    src=self.wip_dir,
+                    dst=self.complete_dir,
+                    ignore=shutil.ignore_patterns("checkpoint.pt"),
+                )
                 print("  \\__Done!")
             except IOError as e:
                 print("  \\__Already exists -- {}".format(e))
@@ -213,48 +252,63 @@ class TrainerOTScratch(object):
 
         # Get experiment's start time
         t0 = time.time()
+        # self.params.max_iter = self.params.max_iter - starting_iter + 1
+        # starting_iter = 0
         # Start training
         iteration = starting_iter
         while iteration <= self.params.max_iter:
             # for iteration in range(starting_iter, self.params.max_iter + 1):
-            for i, (x, y) in enumerate(self.data_loader):
+            for i, (_, factors) in enumerate(self.data_loader):
+                # data = data.squeeze()
+                factors = factors.squeeze()
+                factors = factors.transpose(0, 1)
+                index = torch.randint(0, self.params.num_support_sets, (1, 1), requires_grad=False)
+                data = self.dataset.sequence_by_index(index, factors.cpu().numpy())
+                data = data.squeeze()
+
                 iteration = iteration + 1
                 # Get current iteration's start time
                 iter_t0 = time.time()
 
                 # Set gradients to zero
                 vae_optimizer.zero_grad()
-                support_sets.zero_grad()
-                # reconstructor.zero_grad() # TODO: maybe this is supposed to be generator.zero_grad()??
+                support_sets_optim.zero_grad()
+                reconstructor_optim.zero_grad()
                 if self.use_cuda:
-                    x = x.cuda()
-                x = mnist_color(x)
+                    data = data.cuda()
+                x = data[0]
 
-                index = torch.randint(0, self.params.num_support_sets, (1, 1), requires_grad=False)
-                if hasattr(self.params, "num_support_dipoles"):
-                    half_range = self.params.num_support_dipoles // 2
-                else:
-                    # default sets to max possible value
-                    half_range = mnist_trans.max_step - 1
-                # If you want to have x_0 start with different angle/color/scale
-                # x = mnist_trans(x, torch.randint(0, self.params.num_support_sets,(1,1)), torch.randint(0, half_range//2,(1,1)) )
+                # Generate one-hot index
+                onehot_idx = torch.zeros(x.size(0), self.params.num_support_sets, requires_grad=False)
+                onehot_idx[:, index] = 1.0
+                half_range = self.params.num_support_dipoles // 2
                 recon_x, mean, log_var, z = generator(x)
-                # prior probability
+                # prior distribution
                 std = torch.exp(log_var / 2.0)
                 prob_z = Normal(0.0, 1.0)
                 rho = prob_z.log_prob(z)
-                # posterior probability
+                # posterior distribution
                 prob_zt = Normal(mean, std)
                 rho_t = prob_zt.log_prob(z)
-                vae_loss = self.loss_fn(recon_x, x, mean, log_var)
+                # generate sequence and classify
+                x_seq = x
+                for t in range(1, half_range + 1):
+                    x_t = data[t]
+                    x_seq = torch.cat([x_seq, x_t], dim=1)
+                index_pred = reconstructor(x_seq, iteration)
+                vae_loss = self.loss_fn(recon_x, x, mean, log_var) + self.kl_index(
+                    (index_pred + 1e-20).log(), (onehot_idx + 1e-20)
+                )
+
                 for t in range(1, half_range + 1):
                     with torch.no_grad():
-                        x_t = mnist_trans(x, index, t)
+                        x_t = data[t]
+
                     time_stamp = t * torch.ones(1, 1, requires_grad=True)
+                    energy, loss_pde_tmp, uz, uzz = support_sets.index_forward(index_pred, z, time_stamp)
+                    _, _, _, uzz_prior = prior.index_forward(index_pred, z, time_stamp, rho)
 
-                    energy, loss_pde_tmp, uz, uzz = support_sets(index, z, time_stamp)
-                    _, _, _, uzz_prior = prior(index, z, time_stamp, rho)
-
+                    # Update rho and z
                     rho_t = rho_t - (uzz + 1).abs().log()
                     rho = rho - (uzz_prior + 1).abs().log()
                     z = z + uz
@@ -266,7 +320,6 @@ class TrainerOTScratch(object):
                     img_shifted = generator.inference(z)
 
                     vae_loss += self.bce(img_shifted, x_t)
-
                     if t == 1:
                         loss_pde = loss_pde_tmp
                         rho_loss1 = rho_loss1_tmp
@@ -284,7 +337,11 @@ class TrainerOTScratch(object):
                 )
                 loss.backward()
 
+                # torch.nn.utils.clip_grad_norm_(support_sets.parameters(), 5)
+                # torch.nn.utils.clip_grad_norm_(generator.parameters(), 5)
+                # Perform optimization step (parameter update)
                 support_sets_optim.step()
+                reconstructor_optim.step()
                 vae_optimizer.step()
 
                 # Update tensorboard plots for training statistics
@@ -311,7 +368,7 @@ class TrainerOTScratch(object):
                 if iteration % self.params.log_freq == 0:
                     self.log_progress(iteration, mean_iter_time, elapsed_time, eta)
 
-                # Save checkpoint model file and support_sets model state dicts after current iteration
+                # Save checkpoint model file and support_sets / reconstructor model state dicts after current iteration
                 if iteration % self.params.ckp_freq == 0:
                     # Build checkpoint dict
                     checkpoint_dict = {
@@ -319,6 +376,9 @@ class TrainerOTScratch(object):
                         "support_sets": support_sets.state_dict(),
                         "prior": prior.state_dict(),
                         "vae": generator.state_dict(),
+                        "reconstructor": reconstructor.module.state_dict()
+                        if self.multi_gpu
+                        else reconstructor.state_dict(),
                     }
                     torch.save(checkpoint_dict, self.checkpoint)
         # === End of training loop ===
@@ -326,29 +386,51 @@ class TrainerOTScratch(object):
         # Get experiment's total elapsed time
         elapsed_time = time.time() - t0
 
+        # Save final support sets model
+        support_sets_model_filename = osp.join(self.models_dir, "support_sets.pt")
+        torch.save(support_sets.state_dict(), support_sets_model_filename)
+
+        # Save final shift predictor model
+        reconstructor_model_filename = osp.join(self.models_dir, "reconstructor.pt")
+        torch.save(
+            reconstructor.module.state_dict() if self.multi_gpu else reconstructor.state_dict(),
+            reconstructor_model_filename,
+        )
+
         for _ in range(10):
             print()
         print("#.Training completed -- Total elapsed time: {}.".format(sec2dhms(elapsed_time)))
 
         print("#. Copy {} to {}...".format(self.wip_dir, self.complete_dir))
         try:
-            shutil.copytree(src=self.wip_dir, dst=self.complete_dir, ignore=shutil.ignore_patterns("checkpoint.pt"))
+            shutil.copytree(
+                src=self.wip_dir,
+                dst=self.complete_dir,
+                ignore=shutil.ignore_patterns("checkpoint.pt"),
+            )
             print("  \\__Done!")
         except IOError as e:
             print("  \\__Already exists -- {}".format(e))
 
-    def eval(self, generator, support_sets, prior):
+    def eval(self, generator, support_sets, reconstructor, prior):
         neg_likelihood = []
-        starting_iter = self.get_starting_iteration(support_sets, generator, prior)
+        starting_iter = self.get_starting_iteration(support_sets, reconstructor, generator, prior)
         support_sets.eval()
+        reconstructor.eval()
         generator.eval()
         prior.eval()
         prior_z0 = Normal(0.0, 1.0)
-        for i, (x, y) in enumerate(self.data_loader):
+        iteration = 0
+        for i, (_, factors) in enumerate(self.data_loader):
+            factors = factors.squeeze()
+            factors = factors.transpose(0, 1)
+            index = torch.randint(0, self.params.num_support_sets, (1, 1), requires_grad=False)
+            data = self.dataset.sequence_by_index(index, factors.cpu().numpy())
+            data = data.squeeze()
             with torch.no_grad():
                 if self.use_cuda:
-                    x = x.cuda()
-                    x = mnist_color(x.cuda())
+                    data = data.cuda()
+                x = data[0]
                 recon_x, mean, log_var, z = generator(x)
                 std = torch.exp(log_var / 2.0)
                 q = Normal(mean, std)
@@ -360,60 +442,108 @@ class TrainerOTScratch(object):
                 neg_likelihood.append(neg_logpx_z.sum() / recon_x.size(0))
         print("logpx", sum(neg_likelihood) / len(neg_likelihood))
         neg_likelihood_transformed = []
-        for i, (x, y) in enumerate(self.data_loader):
+        for i, (_, factors) in enumerate(self.data_loader):
+            factors = factors.squeeze()
+            factors = factors.transpose(0, 1)
+            # for index in range(0, 5):
+            index = torch.randint(0, self.params.num_support_sets, (1, 1), requires_grad=False)
             with torch.no_grad():
+                data = self.dataset.sequence_by_index(index, factors.cpu().numpy())
+                data = data.squeeze()
                 if self.use_cuda:
-                    x = x.cuda()
-                    x = mnist_color(x.cuda())
-            for index in range(0, self.params.num_support_sets):
-                with torch.no_grad():
-                    recon_x, mean, log_var, z = generator(x)
-                    q = Normal(mean, std)
-                    log_q_z = q.log_prob(z)
-                    log_p_z = prior_z0.log_prob(z)
-                    # z_prior = z.clone()
-                for t in range(1, self.params.num_support_dipoles // 2):
-                    with torch.no_grad():
-                        x_t = mnist_trans(x, index, t)
-                        _, mean_xt, log_var_xt, _ = generator(x_t)
-                        std_xt = torch.exp(log_var_xt / 2.0)
-                    time_stamp = t * torch.ones(1, 1, requires_grad=True)
-                    _, uz, uzz = support_sets.inference(index, z, time_stamp, mean_xt, std_xt)
-                    _, uz_prior, uzz_prior = prior.inference(index, z, time_stamp)
-                    with torch.no_grad():
-                        log_q_z -= (uzz + 1).abs().log()
-                        log_p_z -= (uzz_prior + 1).abs().log()
-                        z += uz
-                        img_shifted = generator.inference(z)
-                        p = Normal(loc=img_shifted, scale=1.0)
-                        neg_logpx_z = (
-                            -1 * p.log_prob(x_t).flatten(start_dim=1).sum(-1, keepdim=True)
-                            + log_p_z.flatten(start_dim=1).sum(-1, keepdim=True)
-                            - log_q_z.flatten(start_dim=1).sum(-1, keepdim=True)
-                        )
-                        neg_likelihood_transformed.append(neg_logpx_z.sum() / recon_x.size(0))
-        print("logpx transformed", sum(neg_likelihood_transformed) / len(neg_likelihood_transformed))
-        eq_err_all = []
-        for index in range(0, self.params.num_support_sets):
-            for i, (x, y) in enumerate(self.data_loader):
-                with torch.no_grad():
-                    if self.use_cuda:
-                        x = x.cuda()
-                        x = mnist_color(x.cuda())
+                    data = data.cuda()
+                x = data[0]
                 recon_x, mean, log_var, z = generator(x)
-                eq_loss = 0.0
-                for t in range(1, self.params.num_support_dipoles // 2):
-                    with torch.no_grad():
-                        x_t = mnist_trans(x, index, t)
-                        _, mean_xt, log_var_xt, _ = generator(x_t)
-                        std_xt = torch.exp(log_var_xt / 2.0)
-                    _, shift, u_zz = support_sets.inference(
-                        index, z, t * torch.ones(1, 1, requires_grad=True), mean_xt, std_xt
+                std = torch.exp(log_var / 2.0)
+                q = Normal(mean, std)
+                log_q_z = q.log_prob(z)
+                log_p_z = prior_z0.log_prob(z)
+            for t in range(1, self.params.num_support_dipoles // 2):
+                x_t = data[t]
+                time_stamp = t * torch.ones(1, 1, requires_grad=True)
+                _, uz, uzz = support_sets.inference(index, z, time_stamp)
+                _, _, uzz_prior = prior.inference(index, z, time_stamp)
+                with torch.no_grad():
+                    log_q_z -= (uzz + 1).abs().log()
+                    log_p_z -= (uzz_prior + 1).abs().log()
+                    z += uz
+                    img_shifted = generator.inference(z)
+                    p = Normal(loc=img_shifted, scale=1.0)
+                    neg_logpx_z = (
+                        -1 * p.log_prob(x_t).flatten(start_dim=1).sum(-1, keepdim=True)
+                        + log_p_z.flatten(start_dim=1).sum(-1, keepdim=True)
+                        - log_q_z.flatten(start_dim=1).sum(-1, keepdim=True)
                     )
-                    z += shift
-                    with torch.no_grad():
-                        img_shifted = generator.inference(z)
-                    eq_loss += (img_shifted.detach() - x_t).abs().sum() / x_t.size(0)
-                eq_err_all.append(eq_loss)
-            print("eq err", index, sum(eq_err_all) / len(eq_err_all))
+                    neg_likelihood_transformed.append(neg_logpx_z.sum() / recon_x.size(0))
+        print(
+            "logpx transformed",
+            sum(neg_likelihood_transformed) / len(neg_likelihood_transformed),
+        )
+        eq_err_all_traverse0 = []
+        eq_err_all_traverse1 = []
+        eq_err_all_traverse2 = []
+        eq_err_all_traverse3 = []
+        eq_err_all_traverse4 = []
+        # for index in range(0,5):
+        for i, (_, factors) in enumerate(self.data_loader):
+            eq_loss_traverse0 = 0.0
+            eq_loss_traverse1 = 0.0
+            eq_loss_traverse2 = 0.0
+            eq_loss_traverse3 = 0.0
+            eq_loss_traverse4 = 0.0
+            factors = factors.squeeze()
+            factors = factors.transpose(0, 1)
+            index = torch.randint(0, self.params.num_support_sets, (1, 1), requires_grad=False)
+            data = self.dataset.sequence_by_index(index, factors.cpu().numpy())
+            data = data.squeeze()
+            if self.use_cuda:
+                data = data.cuda()
+            x = data[0]
+            x_seq = x
+            with torch.no_grad():
+                for t in range(1, self.params.num_support_dipoles // 2 + 1):
+                    x_t = data[t]
+                    x_seq = torch.cat([x_seq, x_t], dim=1)
+            index_pred = reconstructor(x_seq, iter=100001)
+            recon_x, mean, log_var, z = generator(x)
+            # std = torch.exp(log_var / 2.0)
+            # prob_zt = Normal(mean, std)
+            # eq_loss = 0.0
+            # eq_loss_traverse = 0.0
+            # rho_z = prob_zt.log_prob(z)
+            for t in range(1, self.params.num_support_dipoles // 2 + 1):
+                with torch.no_grad():
+                    x_t = data[t]
+                    recon_xt, _, _, _ = generator(x_t)
+                _, shift, u_zz = support_sets.index_inference(index_pred, z, t * torch.ones(1, 1, requires_grad=True))
+                z += shift
+                # rho_z = rho_z - (u_zz + 1).abs().log()
+                with torch.no_grad():
+                    img_shifted = generator.inference(z)
+                if index == 0:
+                    eq_loss_traverse0 += (img_shifted.detach() - x_t).abs().sum() / x_t.size(0)
+                elif index == 1:
+                    eq_loss_traverse1 += (img_shifted.detach() - x_t).abs().sum() / x_t.size(0)
+                elif index == 2:
+                    eq_loss_traverse2 += (img_shifted.detach() - x_t).abs().sum() / x_t.size(0)
+                elif index == 3:
+                    eq_loss_traverse3 += (img_shifted.detach() - x_t).abs().sum() / x_t.size(0)
+                elif index == 4:
+                    eq_loss_traverse4 += (img_shifted.detach() - x_t).abs().sum() / x_t.size(0)
+            if index == 0:
+                eq_err_all_traverse0.append(eq_loss_traverse0)
+            elif index == 1:
+                eq_err_all_traverse1.append(eq_loss_traverse1)
+            elif index == 2:
+                eq_err_all_traverse2.append(eq_loss_traverse2)
+            elif index == 3:
+                eq_err_all_traverse3.append(eq_loss_traverse3)
+            elif index == 4:
+                eq_err_all_traverse4.append(eq_loss_traverse4)
+        print("eq err traverse 0", sum(eq_err_all_traverse0) / len(eq_err_all_traverse0))
+        print("eq err traverse 1", sum(eq_err_all_traverse1) / len(eq_err_all_traverse1))
+        print("eq err traverse 2", sum(eq_err_all_traverse2) / len(eq_err_all_traverse2))
+        print("eq err traverse 3", sum(eq_err_all_traverse3) / len(eq_err_all_traverse3))
+        print("eq err traverse 4", sum(eq_err_all_traverse4) / len(eq_err_all_traverse4))
+
         return None
